minor improvements
familiarizing myself with this again
want to add typehints
started detecting when you can go up

think i'm leaning towards rewriting this so that by default you start on
a large board and it never grows, but rather you just keep playing. i
think it will be nice to not have to update the board all the time. that
said, it's not _that_ bad.

I almost want a way to represent this game sequentially rather than
spatially. like a linked list of moves, but encoding the rules into that
kind of DS seems much harder.